SurfaceFlinger: Rip out Input caching scheme.
A few weeks ago we implemented a scheme where we won't update
InputWindowInfo if no windows with InputChannels are updated.
However in preparation for an occlusion detection fix we will be
sending input windows for all buffered surfaces (regardless of whether
they have an InputChannel) to InputDispatcher. This means the caching
scheme is effectively useless.

Bug: 152064592
Test: Existing tests pass
Change-Id: I49d0afa9ef6bd620c7632338a5dacca724bd46b8