Allow touch if the new point is not MT_TOOL_PALM
We would cancel all touch when we recevied MT_TOOL_PALM if there are
accidental touches, that could cause the device can't touch anymore.

This ignore these MT_TOOL_PALM points, so if all valid points are
released, we could reset the abort flag and the new valid point
could be treated as new down event.

Test: atest inputflinger_tests
Bug: 137221719
Change-Id: Ib376b912d389c97bc5127fd9b7f4ecb168cd266d