BLASTBufferQueue: Don't use mNextTransaction on Binder thread.
The "mUseNextTransaction" boolean was intended to ensure that we only
used mNextTransaction when calling processBuffer from frameAvailable
and not when processing the queue from the transaction callback. However
this didn't quite work because we can end up returning early from
processNextBuffer() and then the next call (coming from the Transation callback)
will use mNextTransaction from the binder thread. If we randomly access
mNextTransaction from random threads it's very difficult for other threads
to use, so we avoid this directly.

Test: Existing tests pass
Bug: 153120755
Change-Id: Ic261d1fdf3f1458f208face572f2780c1043192c