Pick config based on pixel format.
Currently GLESRenderEngine gets a config if EGL_KHR_no_config_context is
not supported and uses that config to create dummy surfaces if
EGL_KHR_surfaceless_context is not supported. With this patch
GLESRenderEngine can pick a config with a different pixel format rather
than default to 8 bits per channel.

Minor: Remove EGLAttributeVector
Bug: b/145968912
Test: atest librenderengine_test
Change-Id: If242a72be780ad398643b572ae9fab3840b5084d