Trace acquire fence at queue time instead of at latch time
For apps that do not use frame pacing, there is a chance that they keep
dequeueing and queueing a buffer before the SF wake up. In this case,
the older queue item gets dropped by SurfaceFlinger if the expected present
time of the later submitted buffer gets satisfied. The GPU work still
happens for all the submissions however. The acquire fence gets signalled
whenever GPU is done with such work. We currently do not trace this
acquirefence and instead, we trace only the one that gets latched by
flinger. This behavior can break the new phases UI where the slices are
formed based on certain assumptions. This change fixes that by tracing
AcquireFence at queue time instead of latch time.

Test: Take a trace with blur enabled.
Bug: 159472563
Change-Id: I329a11c53dc18eb5d03df096c19dba5c9704ea4a