dalvik sysprop settings are optional
dalvik sysprop settings in these mk files are changed to optional ones
using the new 'a ?= b' syntax, to better express the intent that these
provides safe default values which will be used when they are set in
a product-specific mk file.

This is also in preparation for
I9c073a21c8257987cf2378012cadaeeeb698a4fb, which when lands will
prohibit duplicated sysprop assignments.

Bug: 117892318
Bug: 158735147
Test: m
Change-Id: Ia3528c0c2f9715ff10b85b549d739c9263ab6a7e