Add reconnect method to vibrator HalWrapper
This will allow us to keep the wrapper when the vibrator HAL service
becomes unavailable (and also keep cached capabilities that won't change
for a device). This also makes the connect code in HalController
simpler, since it doesn't need to keep as many boolean flags to avoid
trying to reconnect to unavailable HAL versions.

Bug: b/153418251
Test: atest libvibratorservice_test
Change-Id: Ied50dd85a5fbba33d6bc58f73e626bb6df37b362