Add prefix to transform dump to align the data better
In some cases, we want to add indentations to every line of the
transform dump. Added a prefix argument so the matrix dumped can be
aligned based on the caller.

Also cleaned up how rotation and transformation are logged. If it's one
of the known values, it will just log the string and not the full int
value.

Test: InputDispatcher and InputTransport dumps
Change-Id: I86ed949931fdb233c62349fec9a6749298cccad2