Use ScreenCaptureResults in screenCapture helper functions
It makes more sense to directly send the ScreenCaptureResults object to
the screenCapture helper functions so the render function can populate
the results. This makes it easier when adding new result values since
they likely want to be set for any screenshot path request.

Currently, the result object contains the buffer, whether secure layers
were captured, and the dataspace it was captured in.

Test: SurfaceFlinger_test
Bug: 162367424
Change-Id: I3b1f5264051bb2e885ce0d0df0cd18c8f55afdd1