SF: use parent layer frame rate
If an ancestor of a layer set a frame rate, and that layer doesn't
have a frame rate configured, use the one of the ancestor. The reason
for this is to support the high refresh rate deny list where WM will
set the frame rate on a container layer which is not visible. With this
change we will treat all child layers of that container as if they voted
themselves.

Test: test app that sets preferredDisplayModeId
Bug: 163079696
Change-Id: I3c55d393af72e19cd7b4f107d8cc0b2e85289d96
Merged-In: I3c55d393af72e19cd7b4f107d8cc0b2e85289d96
(cherry picked from commit 20dcde8e35b003a82e4c14d39c76f93a16ec6e9d)