Add back in mBufferId to GraphicBuffer ABI.
Removing mBufferId and getBufferId() changed GraphicBuffer's ABI, which
breaks some GSI builds. It's not well understood where the break
happens, but it's likely due to reinterpret_casting a raw pointer over
a pass-through HAL or similar stable API, which causes issues when
loading GSI as a prebuilt on top of an old Android version. For now,
fixing the ABI break to band-aid the issue

Bug: 160443973
Test: builds
Test: Flash Android 11 GSI on top of Android 10 vendor iage
Test: CtsMediaV2TestCases

Change-Id: Ib6bc78b1651169e9ad4b2602839b84432ffa58c9