Integrate FakeInputReaderContext to the real implement
We use FakeInputReaderContext to mock InputReaderContext and watch if
the interface has been called when we test InputMapperTest and
InputDeviceTest, but it may lost some logic from the real implementation
in InputReader.

This patch will move FakeInputReaderContext into
InstrumentedInputReader and use it when create new device to prevent
getting unexpected results after we modified InputReader.

Bug: 141329037
Test: atest inputflinger_tests
Change-Id: I744908156912116ae85d132873ef8da258e69731